Remove Fixed Input endpoint as it is unused
