+++ test/unit_tests/workflows/test_config_db_code.py
@@ -3,7 +3,7 @@ from more_itertools.recipes import first_true
 from sqlalchemy import not_, select
 from sqlalchemy.orm import joinedload
 
-from orchestrator.api.api_v1.endpoints.fixed_input import fi_configuration
+from orchestrator.utils.fixed_inputs import fi_configuration
 from orchestrator.db import ProductTable, WorkflowTable, db
 from orchestrator.db.models import FixedInputTable
 from orchestrator.services import products
